refactor(shared): lots of helpers
